fix function location
